+++ include/dlaf/factorization/cholesky.h
@@ -11,7 +11,7 @@
 
 #include <blas.hh>
 #include "dlaf/communication/communicator_grid.h"
-#include "dlaf/factorization/cholesky/cholesky_mc.h"
+#include "dlaf/factorization/cholesky/mc.h"
 #include "dlaf/matrix.h"
 #include "dlaf/types.h"
 #include "dlaf/util_matrix.h"
